requirements.txt
# Dev environment
pip
autopep8

# App
flask
mediapipe
supabase
python-dotenv
requests